refactor: moved tests from osc_test to other test file
