+++ client_test.go
@@ -1,1 +1,15 @@
 package osc
+
+import "testing"
+
+func TestClientSetLocalAddr(t *testing.T) {
+	client := NewClient("localhost", 8967)
+	err := client.SetLocalAddr("localhost", 41789)
+	if err != nil {
+		t.Error(err.Error())
+	}
+	expectedAddr := "127.0.0.1:41789"
+	if client.laddr.String() != expectedAddr {
+		t.Errorf("Expected laddr to be %s but was %s", expectedAddr, client.laddr.String())
+	}
+}
